Convert to new override syntax
This is effectively a manual cherry-pick and squash of commits:

4249268041b879b3d2bae338d5de88f4f2d54b26
07f4ab8c2833ae5377b97cc9e3820bed67e539d0

from the master branch with additional review and fixups.

The intent of these changes is to minimize the effort to backport
fixes from the master branch, which has been updated in preparation
for the switch to the next Yocto LTS release in early 2022.

Bug-AGL: SPEC-4144

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I3b47f1b7d62038b314a17763d24a64f6d4e15e72