html5-homescreen: increase wait for attempting to launch homescreen
Reduce chances to fail initialization of HTML5-based UI on boot,
because of trying to launch homescreen while WAM is still not ready
to create applications.

This is a workaround. The right fix will be making WAM actually wait
for processing IPC messages while it is not ready to launch first
application, and making homescreen wait for the initialization of
host service.

Bug-AGL: SPEC-4289
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Id46f87b6ac2d49b0a15d5537b8b70d4ac6e62ef5