Add branch rename workarounds
Add bbappends for ndctl and ttf-noto-emoji recipes to tweak their
SRC_URIs to handle their upstreams renaming the "master" branch to
"main".  This works around the resulting build failures that would
happen otherwise, and patches will be submitted for meta-openembedded
so we can hopefully drop these on the next set of layer updates.

Bug-AGL: SPEC-4361

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ic7756038cffe7bc370f512c835fb1da5954fc387