kuksa-viss-client: add recipe for python3-websockets dependency
The recent addition of the kuksa-viss-client recipe missed adding an
updated recipe for python3-websockets to match the dependency on the
10.x version of websockets that the 0.2.1 version has.  This was
missed during testing with kuksa-dbc-feeder as it seems to not have
the same dependency chain that the kuksa_viss_client tool does.
To allow testing and debugging with kuksa_viss_client, add a backport
of the python3-websockets 10.2 recipe from meta-python's kirkstone
branch.

Bug-AGL: SPEC-4405

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I1c91b2f26c9e451cf9f06005b69f2f8b88d9c318