simple_can_simulator: support different interface for LIN messages
Add a "--lin-interface" option to simple_can_simulator.py to allow
specifying a different CAN interface for the cruise control LIN
messages from the steering wheel in the demo setup.  This allows
pointing at the sllin0 interface in the full demo setup that is
configured with the agl-demo-preload feature.  Previously the
CAN messages for the LIN events were bridged to the primary CAN
interface, but this has proven unreliable in practice.

Bug-AGL: SPEC-4431

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I18b2bb8042569e220c79445d310c620dcebac9fc