Update demo KUKSA.val configuration
In practice mapping multiple CAN signals from the LIN polling to a
smaller number of VIS signals does not work well with the behavior
of the CAN feeder from KUKSA.val when testing on the actual demo
hardware.  To improve the behavior:
- update the VSS schema overlay and DBC feeder mapping configuration
  to use new AGL custom steering wheel switch signals instead of the
  previous signals and mappings.
- add a patch to the DBC feeder to add a per-target duplicate signal
  filtering option, and enable it for all the steering wheel signals
  in the mapping configuration.

The DBC feeder performance issues stemming from synchronous VIS server
updates will be discussed with upstream, but these changes should
yield behavior similar to the previous combination of the low-can and
signal-composer bindings.  With respect to the switch signals, this
should be revisited if implementing a custom LIN signal feeder becomes
an option, as overall it would be better to leverage the standard VSS
schema signals if possible.

Bug-AGL: SPEC-4405

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I702a631d79c4365cb262e9d3a15b9c6be24c8eeb