Disable pulling in meta-python2
Since the meta-python2 layer does not currently support beyond the
Yocto Project kirkstone release, stop pulling it in, and replace the
use of the obsolete pythonnative bbclass with python3native.  This at
least fixes parsing, futher effort will be required to fix building
Chromium and the web apps with Python 3.

Bug-AGL: SPEC-4578

Change-Id: Ib27d111bcae923c5953951b0c40c12c2e30c7464
Signed-off-by: Scott Murray <scott.murray@konsulko.com>