Make cluster dashboard KUKSA.val server configurable
Add CLUSTER_DEMO_VISS_HOSTNAME variables to the cluster-dashboard
and flutter-cluster-dashboard recipes and tweak the installed
configuration files to use the address it specifies.  This provides
a hook for tweaking the configuration for e.g. running in the guest
VMs in the QEMU+KVM demo.  This may need to be tweaked down the
road to use something like multiple packages + the alternatives
mechanism instead.

Bug-AGL: SPEC-4618

Change-Id: I101b2c966fc2923a5f4946ea1c5e95695061c041
Signed-off-by: Scott Murray <scott.murray@konsulko.com>