mpd: Fix patch fuzz warning
The one remaining patch after the kirkstone 4.0.9 upgrade is actually
also no longer required, but git was somehow managing to apply it with
fuzz.  From code inspection the result was some duplicated lines of
code that would still work.  Remove the patch to clean up the fuzz
warning and get us back to using a clean upstream.

Bug-AGL: SPEC-4768

Change-Id: Icb8b8b2fdd847c4fb81a1a7181589b5c8ccaebc0
Signed-off-by: Scott Murray <scott.murray@konsulko.com>