Add base demo images
Some rework to add agl-ivi-demo-base and agl-ivi-demo-base-flutter
images that have all of the required demo services and configuration,
but not our demo UI applications. As part of this, move some of the
conditional package inclusion logic out of the demo platform
packagegroups into the image recipes to make the packagegroup contents
stable, and add applaunchd to the IVI services packagegroup so it'll
always get pulled in. Some of this may still get redone in the future
per SPEC-4813.

Bug-AGL: SPEC-4933

Change-Id: Ief4b7ef3f0720315cbdc579f62b1196c66e237cd
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/29278
Reviewed-by: Lisandro P�rez Meyer <lpmeyer@ics.com>
Reviewed-by: Hiroyuki Ishii <ishii.hiroyuki002@jp.panasonic.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account