[cef] Stop using cros-toolchain
Previous cef and chromium recipes were using chromium's chrome os toolchain as
it was the closest to what were needed to build chromium for target devices.
However it contains additional settings and logic that may conflict with the
cef recipe's compiler settings and change executable paths, so this fix
changes to a custom toolchain that can be set according to AGL needs and
preserve paths set in the recipe.

Bug-AGL: SPEC-4976

Signed-off-by: Roger Zanoni <rzanoni@igalia.com>
Change-Id: Idcb7f6232b8d9989be5daa7ed83930a8b9f02df3
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/29419
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account