cef: use vpython3 for commands called by gclient
The probable cause for SPEC-5062 is that chromium's go routines that
create virtualenvs expect vpython environment variables to be present,
so the creation fails and the scripts could be trying to use the build
host python binary.

This can cause issues on some places, for example in the codecs lookup
that raises an exception when an codec isn't found:

https://github.com/pypa/pip/pull/6311/files

The exception raised on SPEC-5062 logs is "unknown encoding: utf-16le",
because the sysroot pip version expects "utf-16-le" instead.

Bug-AGL: SPEC-5062

Signed-off-by: Roger Zanoni <rzanoni@igalia.com>
Change-Id: I16e817140bf819712a5fbe1b006156db608d43b2
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/29659
ci-image-build: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account