Rework gpsd configuration
To remove the use of the "virtual/gpsd-conf" RPROVIDES as part of
futureproofing for YP 5.0 / Scarthgap, rework the machine-specific
gpsd configuration scheme to based on the gpsd-machine-conf recipe
that is intended for this upstream.  The gpsd-machine-conf recipe
is now unconditionally included in the navigation packagegroup,
but it will be empty on machines except those that add their own
content to it.

Bug-AGL: SPEC-5067

Change-Id: I7f1b555f154fe77c643019dae4bb90354ca2a4b4
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/29691
ci-image-boot-test: Jenkins Job builder account
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account