Rework demo image naming
Rename the demo images to align with the following scheme:

  agl-<ivi|cluster|etc>-<image|demo>-<qt|flutter|etc>(-<crosssdk|preconfigured>)

The intent is to make the naming more consistent in preparation for
switching to preconfigured images as a replacement for the existing
agl-demo-preload, agl-demo-cluster-support, agl-kvm-host-* aglsetup.sh
features.

Additionally, the prototype LXC demo image support has been removed
as it is unused and would need updating for these and upcoming
changes.

Bug-AGL: SPEC-5138

Change-Id: I8d907713b09eb47c3cb0ebd5326ab063b2684ea2
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/29864
ci-image-boot-test: Jenkins Job builder account
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>