vss-agl: Fix engine speed datatype in overlays
The Vehicle.Powertrain.CombustionEngine.Speed signal datatype was
accidentally being overwritten as a float in our VSS overlays for
configuring the CAN provider.  I believe this was an artifact of
some cutting in pasting when the signal was added and upstream
recommended switching away from using OBD signals as the plan is
to deprecate those.  There are some downstream changes required in
the homescreen and dashboard applications to handle this being
fixed.

Bug-AGL: SPEC-5204

Change-Id: I3582dbbef6cf915f30a50804174753da12b905ae
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/30173
Tested-by: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Naoto YAMAGUCHI <naoto.yamaguchi@aisin.co.jp>
ci-image-build: Jenkins Job builder account