demo-flutter/qt: Add agl-shell-activation and window mgmt app
These two applications, (shell-activation and window-management) are
useful to have around in agl-demo as they can used as examples but also
to test out and verify that we haven't introduced any regressions.

The window-managenment-client-grpc was already present in the Qt image
but not in the flutter one while agl-shell-activator wasn't present in
neither of them. There's no dependency for either of them, they can installed
as is.

Change-Id: If83fd4b8e9aba516f818bc61623af9e3e95b7564
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/30718
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account